tests/requirements-test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
coverage==7.3.2
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn==1.3.0
scipy>=1.7.0
biopython>=1.80 